refactor: move NUTs to NUT dir
