+++ test/nuts/metadata/report-mdapi.nut.ts
@@ -11,7 +11,7 @@ import { fileURLToPath } from 'node:url';
 import { SourceTestkit } from '@salesforce/source-testkit';
 import { assert, expect } from 'chai';
 import { RequestStatus } from '@salesforce/source-deploy-retrieve';
-import { DeployResultJson } from '../../../../src/utils/types.js';
+import { DeployResultJson } from '../../../src/utils/types.js';
 
 describe('[project deploy report] NUTs with metadata-dir', () => {
   let testkit: SourceTestkit;
