+++ test/nuts/metadata/report.nut.ts
@@ -11,7 +11,7 @@ import { fileURLToPath } from 'node:url';
 import { SourceTestkit } from '@salesforce/source-testkit';
 import { assert, isObject } from '@salesforce/ts-types';
 import { expect } from 'chai';
-import { DeployResultJson } from '../../../../src/utils/types.js';
+import { DeployResultJson } from '../../../src/utils/types.js';
 
 describe('[project deploy report] NUTs with source-dir', () => {
   let testkit: SourceTestkit;
